Move all test code into KubernetesTestComponents
